[move-prover][refactoring] Implementing data invariants for the v2 compilation scheme.
This adds a new processor `data_invariant_instrumentation.rs` to the v2 pipeline. The processor handles data invariants (not global invariants). It injects assumes/asserts for data invariants into the bytecode for the Pack instruction and for the PackRef/PackRefDeep instructions introduced by the memory model. It also adds assumes for data invariants for parameters of verified functions as well as when data is loaded from resource memory.

This processors does *not* implement spec var related pack/unpack functionality, which we currently plan to drop for v2.

Some bytecode baseline test have been added, as well as existing tests which do not use spec vars migrated.

Closes: #7510