[move-prover] Handle `&mut` parameter threading on boogie instead of bytecode level.
This change came out from a conversation with Shaz. Instead of transforming a function `f(&mut)` from `f(x)` to `x := f(x)` already on bytecode level, we defere this step until Boogie generation, and assume for the bytecode semantics that `&mut` parameter passing semantics is supported. This removes some irregularities in the borrow analysis and simplifies the memory instrumentation further.

Closes: #7795