[move-prover] Revert "Revert "Representing structs as ADTs instead of vectors.""
This reverts commit a0fd2f98a6a2361b7218b228cb8b6cc4677fcbd2 after fixing the underlying issue which caused tests to fail.

The issue was that after moving dep analysis int move-lang, lots of modules are included even though they are never used. This led to a large case distinction for $IsEqual_struct in the new ADT boogie backend, which in turn generated prover timeouts. We are now optimizing the implementation of this function by analyzing which structs are actually needed for a given VC condition.

Closes: #8208