[move-prover] Refactor semantics of `let` in spec blocks.
The semantics of `let` bindings in spec block has been a continued source of confusion. In the old world, for `let x = E`, the name `x` was differently interpreted depending on the context is used in (pre or post state). Also, because of this, `old(..)` could not be used in let bindings.

This PR changes lets to make them more like folks expect: the value of `let` once evaluated is always the same. There are now two lets: `let x = E` is evaluated in the pre-state, and `let post x = E` in the post state, where `E` can use the `old` expression. Moreover, `old(..)` can now be freely used for schema parameters and schema inclusion conditions.

As this change required a lot of changes in the framework specs, this PR also enhances debugging capabilites via the `--trace` flag. Now each member in a spec block is traced, specifically the value of `let` bindings.

The PR also adds additional consistency checks for the correct usage of `old(..)`. This was needed for easier upgrade of the framework to the new let-style. Now applying `old` to an invariant expression creates a type check error, including if this expression is imported into a schema via schema expansion.

Closes: #8326