[stackless-vm] pass an option to loop-to-dag pass to control the emit of Stop or Jump
This partially reverts PR #8295 such that Stop does not need to carry a
funky jump-to Label which is not used in proving. The solution is to add
a `for_interpretation` flag in the ProverOptions and pass option to the
LoopAnalysis pass. If `for_interpretation` is set, emit Jump(label) else
emit Stop.

The reason why the solution in PR #8295 does not work is that the
SpecInstrumentation pass after LoopAnalysis schedules a LiveVarAnalysis
run after the instrumentation. And the LiveVarAnalysis considers Stop as
a terminate instruction instead of a branch instruction, which is
correct in proving mode. But as a result, loop induction variables,
which are supposed to be updated at the end of the loop body, is cleared
out because it is not used past Stop. So, instead of teaching
LiveVarAnalysis to know that Stop is a branch instruction in
interpretation mode, it is better to directly emit the Jump in the
LoopAnalysis pass.

Closes: #8316