[move-prover] Using immutable collections for SetDomain and MapDomain.
This reimplements the collection types for DFA using base types from the [`im' crate`](https://docs.rs/im/15.0.0/im/).

In DFA, we often have types like maps or sets which are copied and refined with each basic block. Traditional Rust collections are not well suited for this, because for each instruction we need to make a deep clone of the input domains. The `im` types are designed with structure sharing in mind. Clone is O(1), and subsequent write leads to structure sharing of old and new
value.

The API of those collections is slightly different which lead to some adoptions in our code. The PR also attempts to complete the wrapping traits for SetDomain and MapDomain such that they (mostly) behave as the underlying representation.

Closes: #8391