[move-cli] test out multi-module publishing via Move CLI first
The `move-cli sandbox publish` command now accepts a new optional
command line argument -m/--override-ordering that allows manual
specification of which compiled modules to be published and in what
order.

Specifying any of the `-m/--override-ordering` flag will invoke the
multi-module publishing flow in the Move VM (compared with the
traditional one-module-at-a-time flow).

For example: `sandbox publish -m A -m B` will prepare a module vector
[A, B] to be sent for publication while `-m B -m A` will create a vector
[B, A]. And depending on the relationship betweem module A and B, it is
possible that only one particular ordering can be allowed. For more
detailed use cases, see the test case `multi_module_publish`.

It is possible to specify less compiles than what gets compiled. For
example, if three modules A, B, C are compiled and only `-m A -m B` is
specified, then only A and B will be sent to Move VM for publication.

It is also possiible to specify one module multiple times, e.g.,
`-m A -m A`. The Move CLI will not complain about it but the Move VM
will as it checks for no duplication in the vector of modules.

It is, however, not possible to specify a non-compiled module (e.g.,
module D which is not compiled) for publication. An error will be raised
in this case.

Closes: #8555