[move-prover] Tech report: global invariants first take
This adds a draft section about global invariants in general, and the specific problem of genericity and monomorphization.

Writing this gave some interesting insights which might be relevant for the ongoing work on generic invariants. Specifically, the problem of monomorphization should be seen independently of that of invariant injection.

This does not yet tackle modular verification and suspension of invariants.

Closes: #8841