[verifying-client] Add FileStateStore to store TrustedStates in local FS
+ `FileStateStore` is a `StateStore` that stores `TrustedState`s in files
inside a given directory on the local filesystem.

+ In order to provide atomicity guarantees, we have two files on-disk
where the highest version state file is durable and not being written
to at all times. When a writer wants to store a new state, it gets an
exclusive lock and writes to the on-disk state file with the lower
version (if the new state is actually newer). That way, if the machine
crashes or restarts during a store, we'll only corrupt the older state
file (the one being written) and the newer state file will remain untouched.

+ Note: it only supports either one writer xor one reader at a time.