[move-cli] allow tests to run in a tempdir-based workspace
We have been experiencing flakiness in the move-cli tests recently,
as observed in #8917 and #8846 (although they are very rare and very
hard to reproduce locally).

The reason is some race condition between the datatest directory
scanning and the dynamically created `build` and `storage` directories
during the running of a test.

Following are some facts that should help illustrate the problem:
- Almost all Move CLI tests are datatests.

- When running a test in `tests/A`, we will create two directories,
  `tests/A/build` and `tests/A/storage`, and subsequently remove them
  when test `A` finishes.

- When running test `B`, we will invoke the `datatest::runner` and one
  thing the runner will do is to `WalkDir` over directories and files
  under `tests`, which include `tests/A`, `tests/B`, and if A and B
  runs at the same time, `tests/A/build` and `tests/A/storage`.

Now, with any luck (or bad luck), the following race condition may be
triggered
- Test A fires up and creates `tests/A/build`
- Test B fires up and the `WalkDir` noticed the existence of directory
  `tests/A/build`,
- Test A finishes, and removed `tests/A/build`
- Test B now gets confused and panicked, because `tests/A/build` no
  longer exists anymore.

Judged by the abort trace, this is what happened in the two issues
flagged by the flaky bot.

The fix is to redirect the creation of `build` and `storage`
directories outside of the path covered by the `datatest::harness`,
at least for CI tests.

This commit implements such a fix by redirecting the workspace dir
for a test execution into a `tempfile::tempdir()`.`

Closes: #8938