[move-prover] refactor the verification analysis
The verification analysis has three tasks:
1. calculate the set of global invariants that are applicable to a
   function --- the "working-set" of invariants in this function.
2. based on the invariant applicability info, derive which function
   should be marked as verified (and subsequently, inlined).

With the suspendable global invariant feature, one pre-requisite
for task 1 is to derive two sets of functions based on several
invariant-related pragmas:
- the E-set: a suspendable invariant will not be verified in the
  function body but will be assumed at the function entry and asserted
  at function return.
- the N-set: a suspendable invariant might not hold in any code position
  in the function, therefore, the invariant can never be assumed in the
  function, neither can we assert the invariant.

If there are some errors in the pragmas, those errors will be reported
in this phase too. Such errors include:
- external-facing functions are not allowed in the N set (i.e., have
  invariant checking completely turned-off), UNLESS they don't modify
  any memory that are checked in any suspendable invariant.
- a function cannot be both on the E set and N set, i.e., a function
  cannot have invariant checking turned-off completely while also
  checking the invariant at the function return.

With the information on the E-set and N-set, we can calculate the
"working-set" of invariants per each function. An invariant `I` is
considered to be applicable to function `F` iff there exists
- some memory of type `M1` that is mentioned in `I`, and
- some memory of type `M2` that is used in `F`, and
- `M1` unifies with `M2`.

Furthermore, depending on how `M2` is used in `F`, there are four
sub-cases:
- `M2` is accessed (read or written) directly in `F`
- `M2` is accessed (read or written) indirectly by a function `F` calls
- `M2` is modified (written) directly in `F`
- `M2` is modified (written) indirectly by a function `F` calls
This suggests that `I` can be relevant to `F` in four ways, depending on
which case `M2` belongs to.

This means that for each function, we can derive four sets of invariants
that are related to the function, casing by how `M2` is used by `F`:
- `I_access`
- `I_modify`
- `I_direct_access`
- `I_direct_modify`

With this few, the invariant suspension and deferrment can be modeled by
operations over these sets: if `F` calls a function `C` which are in the
N-set (i.e., have invariant checking completely turned-off for suspendable
invariants), then what will happen is that all the suspendable invariants
from the sets in `C` (e.g., `I_modify(C)`) will be taken out in the sets
in `C` and subsequently be backpropogated to the caller, which is `F`.
Essentially, `I_direct_access(F)` is updated with the suspendable invariants
in I_access(C)`. And similarly for `I_direct_modify(F)`.

These information will not be useful in the verification analysis phase,
but will be used in the global invariant instrumentation phase.

Last point: this refactored `verification_analysis.rs` lives alongside
with the current version which is `verification_analysis_v2.rs`. The
plan is to not enable it until we have a compatible global invariant
instrumentation pass that can digest the info created in this pass.