[move-prover] an instrumentation pass for the global invariants
An instrumentation pass that consumes the information produced by the
global invariant analysis pass and instrument the global invariants into
the function.

The instrumenter supports two mode of operations, depending on whether
the prover backend supports monomorphization or not:
1) With the option `--boogie-poly`, the instrumenter will instrument
  *instantiated* invariants in the *generic* function (and the generic
  instance is the only function instance).
2) Without the `--boogie-poly` option, the instrumenter will instrument
  *instantiated* invariants *per each instantiation* of a generic
  function (this is the traditional workflow). And this means that a
  function will have multiple instances for verification.

This is not exactly the plan we had before (and does not clearly adhere
to the paper). The original plan is to go for option 1 first and defer
the instantiation of functions to the mono pass. Therefore, the option 2
here is essentially a combination of
1) instrumentation,
2) monomorphization, and
3) optimization (eliminating redundant expressions)

But option 2 does not completely solve the "type-dependent" code problem
because the (`move_to<T>`, `move_to<u64>`) case still requires a second
step of function instantiation and still requires the mono pass to
perform such instantiation.

The main reason why we still have option 2 (and not only that, we made
option 2 as default as of now) is three-fold:
1) I am uncertain of Boogie's monomorphization implementation matches
   the complexity of what we have done here.
2) I want to get at least the whole Diem Framework verifying again to
   test out the whole transformation pipeline, in order to boost our
   confidence.
3) Solving the `move_to<T>; move_to<u64>;` problem requires more than
   instantiation; we need spec language support as well to express the
   fact that the function will surely abort if `T == u64` and may not
   abort otherwise.

With this final piece, we complete the new invariant instrumentation
pipeline (modulus the misaligned implementation plan mentioned above)
and the next PR will switch the pipeline into the new one and fix the
specs in the Diem Framework.

Closes: #9077