[storage] state_store: get_account_count()
This implemented serialization of the new jelly fish internal node format for the state DB to utilize `JellyfishMerkleTree::get_leaf_count()`. As implemented in this commit, if a StateDB is created from scratch, leaf counts will be available on all internal nodes. Sophisticated migration control will be implemented as a follow up.