[mempool] Fix transaction rebroadcast bug
Mempool was assuming that Validators that are no longer connected have
previously broadcasted transactions.  This is problematic because it
assumes the node maintains its memory through the disconnect which is
not necessarily the case.  Additionally, it made it such that we kept
scheduling broadcasts forever for nodes that no longer exist or are
no longer connected.