[move-prover] refactor the way entrypoint assumptions are collected
This is motivated by the fix for uninstantiated invariants. But in
general, we want to maintain a property that if an invariant I is
asserted later in the function, it should be assumed at the entrypoint
of the function. This includes generic invariants, so if I<T = u64> is
assumed in the code, I<T = u64> needs to be asserted in the beginning.

Prior to this commit, entrypoint assumptions are collected by:
```
for each M in F.accessed_memory.all:
    for each I in F.accessed_invariants.all:
        I_with_inst = unify_and_instantiate(M, I)
        add I_with_inst to the entrypoint assumptions
```

This commit splits the collection into two parts:
```
for each B in F.bytecode:
    m_ro = B.get_accessed_but_not_modified_memory
    m_rw = B.get_modified_memory

    for each M in m_ro:
        for each I in F.accessed_invariants.all:
            I_with_inst = unify_and_instantiate(M, I)
            // --> it is fine if I cannot be instantiated
            //     we know that I will not be asserted
            add I_with_inst to the entrypoint assumptions

    for each M in m_rw:
        for each I in F.modified_invariants.direct:
            I_with_inst = unify_and_instantiate(M, I)
            // --> it will not be fine if I cannot be instantiated
            //     which will be implemented in the next commit
            add I_with_inst to the entrypoint assumptions
```

This essentially ensures that whatever is asserted later in the code
matches with whatever is assumed at the beginning of the function.