[move] Bump VS Code test timeout
The GitHub action to test the move-analyzer VS Code extension (the
language server client) would sometimes timeout. When run locally, the
tests spend most of their time booting up an instance of VS Code. Bump
the timeout up 2.5x, on the assumption that this is also why the GitHub
action is taking a while.

Closes: #9615