[move] Remove empty `;` statement from Move IR
The documentation comment for the `EmptyStatement` claims it eases
parsing to allow for an empty statement `;` in Move IR, but I think that
it actually makes things more difficult: the statement `assert(e_1, e_2);`
was being parsed incorrectly for some time, as an assert statement
followed by the empty statement `;`.

I could see how it may make sense to allow for empty statements in a
user-facing programming language such as Move, but for a language we
primarily use to test Move, I think the empty statement should either
lower to a "noop" bytecode instruction, or it should be treated as an error.
And seeing as how there is no "noop" bytecode instruction, I think empty
`;` should result in a parser error, as it does in this commit.

I'm currently working on eliminating structured control flow statements
in Move IR in a separate commit, and removing the empty statement `;`
now would allow me to greatly simplify statements overall (potentially,
I could eliminate the `Statement` enum, because all `Statement` would be
`CommandStatement`).

Closes: #9759