[move-prover] better architecture for simplification pipeline
This creates a new trait, `SpecRewriter` that handles the
transformations on specs.

The `SpecRewriter` has three functions to be implemented by each
individual rewriter:
- `rewrite_module_spec`
- `rewrite_function_spec`
- `rewrite_inline_spec`

Multiple rewriters that implement the `SpecRewriter` trait can be
composed into a new rewriter that pipelines them. This logic is captured
in `SpecRewriterPipeline`, which also implements `SpecRewriter`.

A `SpecRewriter` has an option to write-back the newly generated spec
into a `&mut GlobalEnv` with the `override_with_rewrite` function. But
this is optional and only available in cases where a `&mut GlobalEnv`
can be obtained (e.g., during model building or right after the model
is built). If writing-back is not needed, the three `rewrite_*` funciton
should be sufficient.

Closes: #9821